requirements.txt
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
xformers==0.0.28.post3
onnx==1.17.0
onnxruntime-gpu==1.19.2
deepspeed==0.15.2
accelerate==1.1.1
diffusers==0.31.0
transformers==4.46.3
tokenizers==0.20.1
peft==0.12.0
decord==0.6.0
sentencepiece==0.2.0
opencv-python==4.10.0.84
pyfacer==0.0.4
numpy==1.26.4
numba==0.60.0
insightface==0.7.3
huggingface-hub==0.26.1
facexlib==0.3.0
timm==1.0.9
func_timeout==4.3.5
tensorboard==2.17.1
gradio==5.6.0
spaces==0.30.4
pillow==10.4.0
spandrel==0.4.0
scikit-video==1.1.11
moviepy
wandb
imageio-ffmpeg
ftfy
Jinja2
einops
nvitop